services/encoder_service/requirements.txt
python-dotenv
sentence-transformers
numpy
redis[asyncio]
